Implement basic functionality
Lizard-bot won't respond at all to a post by someone with \#NoBots in
their profile, and tags are restricted to people without \#NoBots.

We can make a basic guess as to which status we are being tagged to
analyze with analyze-ancestor-p, but it would be nice to have a way to
allow whoever tags us to make the decision themselves.

Currently the bot will get notified multiple times for a single
notification, leading to multiple responses. This could be
accomplished with idempotency tokens or a record of statuses we've
already replied to, or both. (idempotency keys have been a FIXME in
tooter for two years, so we'll see)